+++ plugins/caspar/app/utils/asset.unit.test.js
@@ -1,4 +1,4 @@
-import { calculateDurationMs } from './asset.js'
+import { calculateDurationMs, frameRateFractionToDecimal, frameRateFractionToDecimalRounded } from './asset.js'
 
 test('calculateDurationMs should return 0 for 0 duration (string or number)', () => {
   // Test for both string and number '0'
@@ -52,4 +52,36 @@ test('calculateDurationMs should return 0 for STILL image', () => {
     duration: '100'
   }
   expect(calculateDurationMs(item)).toEqual(0)
+})
+
+// frameRateFractionToDecimal tests
+
+test('frameRateFractionToDecimal should return correct value for even frame rates', () => {
+  expect(frameRateFractionToDecimal('1/24')).toEqual(24)
+  expect(frameRateFractionToDecimal('1/25')).toEqual(25)
+  expect(frameRateFractionToDecimal('1/30')).toEqual(30)
+  expect(frameRateFractionToDecimal('1/50')).toEqual(50)
+  expect(frameRateFractionToDecimal('1/60')).toEqual(60)
+})
+
+test('frameRateFractionToDecimal should return correct value for odd (NTSC) frame rates', () => {
+  expect(frameRateFractionToDecimal('1001/24000')).toBeCloseTo(23.976, 3)
+  expect(frameRateFractionToDecimal('1001/30000')).toBeCloseTo(29.97, 2)
+  expect(frameRateFractionToDecimal('1001/60000')).toBeCloseTo(59.94, 2)
+})
+
+test('frameRateFractionToDecimal should return undefined for non-numeric input', () => {
+  expect(frameRateFractionToDecimal('abc/def')).toBeUndefined()
+  expect(frameRateFractionToDecimal('foo')).toBeUndefined()
+  expect(frameRateFractionToDecimal('')).toBeUndefined()
+})
+
+test('frameRateFractionToDecimal should return undefined when divisor is zero or negative', () => {
+  expect(frameRateFractionToDecimal('0/25')).toBeUndefined()
+  expect(frameRateFractionToDecimal('-1/25')).toBeUndefined()
+})
+
+test('frameRateFractionToDecimal should handle non-string input by coercing to string', () => {
+  expect(frameRateFractionToDecimal(undefined)).toBeUndefined()
+  expect(frameRateFractionToDecimal(null)).toBeUndefined()
 })